% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingCitations.R
\name{checkMissingCitations}
\alias{checkMissingCitations}
\title{Check for missing bibliographic entries}
\usage{
checkMissingCitations(entries, bib)
}
\arguments{
\item{entries}{\itemize{
\item missing citation entries to look for, or path to file with such
}}

\item{bib}{\itemize{
\item character vector with bib text, or path to bib file
}}
}
\value{
vector of missing entries
}
\description{
Function to check for missing bibliographic entries in a .bib file.
}
