% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tblr_to_kbl.R
\name{convert_tblr_to_kbl}
\alias{convert_tblr_to_kbl}
\title{Convert \link{tables} \code{tabular} object to \link{knitr} \code{kable} object}
\usage{
convert_tblr_to_kbl(
  tblr,
  col_spec,
  isHTM = NULL,
  isPDF = NULL,
  scale_down = FALSE,
  ltx_font_size = NULL,
  replaceNAs = NULL,
  adjColSpacing = 0
)
}
\arguments{
\item{tblr}{\itemize{
\item \link{tables} \code{tabular} object to convert
}}

\item{col_spec}{\itemize{
\item vector of column ids to apply "border_right" to using \code{\link[kableExtra:column_spec]{kableExtra::column_spec()}}
}}

\item{isHTM}{\itemize{
\item flag indicating whether output is for html (any value other than FALSE indicates TRUE; default=NULL)
}}

\item{isPDF}{\itemize{
\item flag indicating whether output is for pdf (default=NULL; NULL indicates FALSE)
}}

\item{scale_down}{\itemize{
\item flag to scale font size down to fit table within margins (works only for pdf, i.e. \code{isHTM=FALSE})
}}

\item{ltx_font_size}{\itemize{
\item font size for latex table (NULL gives default font size)
}}

\item{replaceNAs}{\itemize{
\item string to replace NAs with (default=NULL does nothing)
}}

\item{adjColSpacing}{\itemize{
\item value to adjust latex column spacing by (default=0; default column spacing = 6)
}}
}
\value{
\link{knitr} \code{kable} object
}
\description{
Function to convert \link{tables} \code{tabular} object to \link{knitr} \code{kable} object.
}
\details{
If \code{isHTM} and \code{isPDF} are both NULL (the defaults), the conversion will be to html.
}
