% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQMD_FromTemplate.R
\name{createQMD_FromTemplate}
\alias{createQMD_FromTemplate}
\title{Create a Quarto markdown document from a template}
\usage{
createQMD_FromTemplate(
  fn,
  ltx = system.file("files/ltx_ExtraLatexIncludes.tex", package = "wtsQMD")
)
}
\arguments{
\item{fn}{\itemize{
\item filename for output QMD file
}}

\item{ltx}{\itemize{
\item filename (full path) for latex include file (see \link{@details})
}}
}
\value{
nothing (invisible(NULL))
}
\description{
Function to create a Quarto markdown document from a template.
}
\details{
Writes contents of \code{system.file("file/qmd_chapter_template.qmd",package="wtsQMD")}
to the desired file. File extension should be '.qmd'. As a default, the QMD will use the file
given by \code{system.file("files/ltx_ExtraLatexIncludes.tex",package="wtsQMD")} as the latex
'include' file.
}
