% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doKbl.R
\name{doKbl}
\alias{doKbl}
\title{Convert dataframe object to \link{knitr} \code{kable} object}
\usage{
doKbl(
  dfr,
  align = NULL,
  border_left = NULL,
  border_right = NULL,
  booktabs = TRUE,
  longtable = TRUE,
  replaceNAs = NULL,
  position = "h",
  digits = getOption("digits"),
  format.args = list(big.mark = ","),
  bootstrap_options = c("striped", "condensed"),
  latex_options = c("striped", "repeat_header"),
  ltx_font_size = NULL,
  scale_down = TRUE,
  adjColSpacing = 0
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe to convert
}}

\item{align}{\itemize{
\item optional vector specifying column alignment (use 'l','c','r': see \code{\link[kableExtra:kbl]{kableExtra::kbl()}})
}}

\item{border_left}{\itemize{
\item vector of column ids to apply "border_left" to using \code{\link[kableExtra:column_spec]{kableExtra::column_spec()}}
}}

\item{border_right}{\itemize{
\item vector of column ids to apply "border_right" to using \code{\link[kableExtra:column_spec]{kableExtra::column_spec()}}
}}

\item{booktabs}{\itemize{
\item flag to use booktabs format
}}

\item{longtable}{\itemize{
\item flag to use longtable format (latex output only)
}}

\item{replaceNAs}{\itemize{
\item string to replace NAs with (default=NULL does nothing)
}}

\item{position}{\itemize{
\item character indicating position (latex output only: "h", "t","p")
}}

\item{format.args}{\itemize{
\item passed on to \code{\link[kableExtra:kbl]{kableExtra::kbl()}}
}}

\item{ltx_font_size}{\itemize{
\item font size for latex table (NULL gives default font size)
}}

\item{scale_down}{\itemize{
\item flag to scale font size down to fit table within margins (latex output only)
}}

\item{adjColSpacing}{\itemize{
\item value to adjust latex column spacing by (default=0; default column spacing = 6)
}}

\item{col_spec}{\itemize{
\item vector of column ids to apply "border_right" to using \code{\link[kableExtra:column_spec]{kableExtra::column_spec()}}
}}

\item{digitis}{\itemize{
\item max number of digits for numeric columns. Can also be a vector of length ncol(dfr).
}}
}
\value{
\link{knitr} \code{kable} object
}
\description{
Function to convert dataframe object to \link{knitr} \code{kable} object.
}
\details{
The conversion to html or latex is determined by \code{\link[knitr:output_type]{knitr::is_latex_output()}}.
}
