% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_TableFunctions.Params.R
\name{doTable.ParamVals}
\alias{doTable.ParamVals}
\title{Create a tabular object for parameter values for TCSAM02 models}
\usage{
doTable.ParamVals(
  dfr,
  ctgs = NULL,
  prcs = NULL,
  lbls = NULL,
  nams = NULL,
  grep_names = NULL,
  param_type = "number"
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe from call to \code{\link[rCompTCMs:extractMDFR.Results.ParameterValues]{rCompTCMs::extractMDFR.Results.ParameterValues()}}
}}

\item{ctgs}{\itemize{
\item parameter categories to select  (default=NULL; i.e., all)
}}

\item{prcs}{\itemize{
\item parameter processes to select   (default=NULL; i.e., all)
}}

\item{lbls}{\itemize{
\item parameter labels to select      (default=NULL; i.e., all)
}}

\item{nams}{\itemize{
\item parameter names to select        (default=NULL; i.e., all)
}}

\item{grep_names}{\itemize{
\item flag to select parameter names using grep
}}

\item{param_type}{\itemize{
\item string ("number" or "vector")
}}
}
\value{
tabular object
}
\description{
Function to create a tabular object for parameter values for TCSAM02 models.
}
