% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_TableFunctions.Params.R
\name{doTable.PsAtBs}
\alias{doTable.PsAtBs}
\title{Create a tabular object for parameters-at-bounds for TCSAM02 models}
\usage{
doTable.PsAtBs(dfr)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe from call to \code{\link[rCompTCMs:extractMDFR.Results.ParametersAtBounds]{rCompTCMs::extractMDFR.Results.ParametersAtBounds()}}
}}
}
\value{
tabular object
}
\description{
Function to create a tabular object for parameters-at-bounds for TCSAM02 models.
}
