% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escapeChars.R
\name{escapeForLatex}
\alias{escapeForLatex}
\title{Escape special characters in a character string for latex.}
\usage{
escapeForLatex(x)
}
\arguments{
\item{txt}{\itemize{
\item character string in which to escape special characters
}}
}
\value{
character vector with special characters escaped.
}
\description{
Function to escape special characters in a character string for latex.
}
\details{
Currently escapes the following characters: "&", "\%", "_", and "'"

"'" is escaped by substituting "\\textquotesingle " from the latex \code{textcomp} package.
}
