% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escapeChars.R
\name{escapeQMD}
\alias{escapeQMD}
\title{Escape strings for QMD processing}
\usage{
escapeQMD(str)
}
\arguments{
\item{str}{\itemize{
\item string to escape
}}
}
\value{
string with special QMD characters escaped with "\\" so \link{cat}'ed result
has properly-escaped character.
}
\description{
Function to escape strings for QMD processing.
}
\details{
The following characters will be escaped:
\itemize{
\item{' - single quote}
\item{" - double quote}
}
}
