% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFunctions.R
\name{findFigLabels}
\alias{findFigLabels}
\title{Find figure labels (in order of first appearance)}
\usage{
findFigLabels(txt, fn = "r_reorderFigs.csv")
}
\arguments{
\item{txt}{\itemize{
\item text to search, or file path for text to search
}}

\item{fn}{\itemize{
\item filename to write labels to
}}
}
\value{
\itemize{
\item tibble with figure labels in orser of first occurrence
}
}
\description{
Function to find figure labels (in order of first appearance).
}
