% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFunctions.R
\name{findTblLabels}
\alias{findTblLabels}
\title{Find table labels (in order of first appearance)}
\usage{
findTblLabels(txt, fn = "r_reorderTbls.csv")
}
\arguments{
\item{txt}{\itemize{
\item text to search, or file path for text to search
}}

\item{fn}{\itemize{
\item filename to write labels to
}}
}
\value{
\itemize{
\item tibble with table labels in order of first occurrence
}
}
\description{
Function to find table labels (in order of first appearance).
}
