% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getFigCaption}
\alias{getFigCaption}
\title{Get a Quarto/knitr chunk figure caption}
\usage{
getFigCaption(xtra = NULL)
}
\arguments{
\item{xtra}{\itemize{
\item extra string to  add to caption
}}
}
\value{
a string
}
\description{
Function to get a Quarto/knitr chunk figure caption.
}
\details{
If the function is not run within the context of a knitr chunk,
then 'Figure' is used as the default caption.
}
