% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getFigFN}
\alias{getFigFN}
\title{Create an appropriate filename for a figure}
\usage{
getFigFN(xtra = NULL)
}
\arguments{
\item{xtra}{\itemize{
\item extra string to add to default filename
}}
}
\value{
string
}
\description{
Function to create an appropriate filename for a figure.
}
\details{
Non-alphanumeric characters in \code{xtra} that cannot be in a label are replaced by "-".
}
