% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileFunctions.R
\name{getImageFN}
\alias{getImageFN}
\title{Get file name with correct extension for an image in the current knitr environment}
\usage{
getImageFN(fn)
}
\arguments{
\item{fn}{\itemize{
\item file name for an image
}}
}
\value{
file name, posibly updated with  extension
}
\description{
Function to get file with correct extension for an image in the current knitr environment.
}
\details{
file extension is converted to ".pdf" in a latex environment or "png" in a
html environment.
}
