% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getLabel}
\alias{getLabel}
\title{Create a Quarto/knitr chunk label}
\usage{
getLabel(xtra = NULL)
}
\arguments{
\item{xtra}{\itemize{
\item extra string to  add to default label
}}
}
\value{
a string
}
\description{
Function to create a Quarto/knitr chunk label.
}
\details{
If the function is not run within the context of a knitr chunk,
then 'lbl' is used as the default label. Non-alphanumeric characters in \code{xtra}
are replaced by "-" in order to conform to Quarto labels
}
