% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{getModelsList}
\alias{getModelsList}
\title{Create a string of model names}
\usage{
getModelsList(mdls)
}
\arguments{
\item{mdls}{\itemize{
\item vector of model names
}}
}
\value{
\itemize{
\item character string of model names
}
}
\description{
Function to create a string of model names.
}
\details{
\itemize{
\item Collapses model names into a string suitable to
include in a sentence by adding commas and "and" appropriately.
}
}
