% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{getOverfished}
\alias{getOverfished}
\title{Determine whether stock is overfished based on stock status value}
\usage{
getOverfished(status)
}
\arguments{
\item{status}{\itemize{
\item stock status value (i.e., B/$B_{MSY}$)
}}
}
\value{
string with overfished status ("overfished", "not overfished")
}
\description{
Function to whether stock is overfished based on stock status value.
}
\details{
overfished designation is determined from status by:
}
