% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{getTierLevel}
\alias{getTierLevel}
\title{Determine Tier level from stock status value}
\usage{
getTierLevel(status)
}
\arguments{
\item{status}{\itemize{
\item stock status value (i.e., B/$B_{MSY}$)
}}
}
\value{
string with tier level ("a", "b", or "c")
}
\description{
Function to determine Tier level from stock status value.
}
\details{
Tier level is determined from status by:
}
