% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasBibEntry.R
\name{hasBibEntry}
\alias{hasBibEntry}
\title{Check for a bibliographic entry}
\usage{
hasBibEntry(entry, bib)
}
\arguments{
\item{entry}{\itemize{
\item citation entry to look for
}}

\item{bib}{\itemize{
\item character vector with bib text, or path to bib file
}}
}
\value{
number of entries found
}
\description{
Function to check for a bibliographic entry in a .bib file.
}
