% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageAsKable.R
\name{imageAsKable}
\alias{imageAsKable}
\title{Allow an image to be printed as if a table (i.e., a kable object)}
\usage{
imageAsKable(fn)
}
\arguments{
\item{fn}{\itemize{
\item filename of image to use as if a table
}}
}
\value{
A character string if output is pdf, otherwise a kable object
}
\description{
function to allow image to be printed as if a table (i.e. a kable object)
}
\details{
This function allows a previously-created image to be printed as a kable object
in a Quarto (qmd) document.
}
