% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertFigInHTML.R
\name{insertFigInHTML}
\alias{insertFigInHTML}
\title{Insert figure into a QMD markdown document for HTML}
\usage{
insertFigInHTML(lst)
}
\arguments{
\item{lst}{\itemize{
\item list with figure info (see \link{@details})
}}
}
\value{
\itemize{
\item invisible NULL
}
}
\description{
Function to insert a figure into a QMD markdown document for HTML.
}
\details{
The input \code{lst} should have elements
\itemize{
\item{pth - path relative to \code{root} to figures folder}
\item{lbl - figure label--must start with 'fig-'}
\item{wid - figure width in inches, optional}
\item{hgt - figure height in inches, optional}
\item{dpi - dots per inch}
}
}
