% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertFlextableIntoQMD.R
\name{insertFlextableIntoQMD}
\alias{insertFlextableIntoQMD}
\title{Insert (print) a \link{flextable} \code{flextable} object into a QMD (Quarto markdown) document}
\usage{
insertFlextableIntoQMD(flx, lbl, cap = lbl, ori = "L")
}
\arguments{
\item{flx}{\itemize{
\item \link{flextable} \code{flextable} object
}}

\item{lbl}{\itemize{
\item table label (i.e., of the form "tbl-..." for cross-references)
}}

\item{cap}{\itemize{
\item table caption
}}

\item{ori}{\itemize{
\item orientation ("L" or "P")
}}
}
\value{
a list wrapped in a list if output environment is latex, otherwise NULL (see \link{details})
}
\description{
Function to insert (print) a \link{flextable} \code{flextable} object into a QMD (Quarto markdown) document.
}
\details{
If the output environment is \emph{not} latex, the resulting table will be printed "here" in the text.
To do this, a QMD cell with table caption is printed via \link{cat}
to the output environment (typically an intermediate markdown file used in the
process of creating the final output). The value returned by the function is NULL.

If the output environment is latex, then a list is returned that should be added to \code{lstTbls}
to be printed in the Tables section. If \code{lst} is the returned list, there are two ways to do this:
\itemize{
\item{\code{lstTbls = c(lstTbls,lst);}}
\item{\code{lstTbls[[lbl]] = lst[[1]];}}
}
where \code{lbl} in the latter has the same value as the input parameter \code{lbl}.

The following will work in either latex or html output environments:\cr
\code{  lstTbls = c(lstTbls,insertFlexTableIntoQMD(flx,lbl,cap,ori));}
}
