% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertImageIntoHTML.R
\name{insertImageIntoHTML}
\alias{insertImageIntoHTML}
\title{Insert an image from a file into a QMD markdown document for HTML}
\usage{
insertImageIntoHTML(
  lst = NULL,
  pth = NULL,
  cap = NULL,
  wid = NULL,
  hgt = NULL,
  dpi = NULL
)
}
\arguments{
\item{lst}{\itemize{
\item list with figure info, or NULL if remaining inputs are given (see \link{@details})
}}

\item{pth}{\itemize{
\item path to image file--ignored if \code{lst} is given
}}

\item{cap}{\itemize{
\item caption (or NULL)--ignored if \code{lst} is given
}}

\item{wid}{\itemize{
\item image width (in inches, or NULL)--ignored if \code{lst} is given
}}

\item{hgt}{\itemize{
\item image width (in inches, or NULL)--ignored if \code{lst} is given
}}

\item{dpi}{\itemize{
\item image dots per inch (or NULL; required if \code{wid} or \code{hgt} is given)--ignored if \code{lst} is given
}}
}
\value{
\itemize{
\item invisible NULL
}
}
\description{
Function to insert an image from a file into a QMD markdown document for HTML.
}
\details{
Intended for use in a QMD rendered in a HTML environment.
The necessary HTML code to insert an image into, e.g., a HTML document
is created and \link{cat}'ed. The QMD chunk that employs this function should
set \code{results="asis"} in the chunk header information.

The corresponding function for insertion into latex is \link{insertImageIntoLatex}.

The input \code{lst} should have the following elements:
\itemize{
\item{pth - path to image relative to \code{root} of document to be created}
\item{lbl - figure label--must start with 'fig-'}
\item{wid - figure width in inches, optional}
\item{hgt - figure height in inches, optional}
\item{dpi - dots per inch (required if \code{wid} or \code{hgt} is given)}
}
}
