% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertImageIntoLatex.R
\name{insertImageIntoLatex}
\alias{insertImageIntoLatex}
\title{Create and paste the latex code to insert an image into a latex document}
\usage{
insertImageIntoLatex(
  pth,
  cap,
  lbl = cap,
  width = 6.5,
  height = 6,
  align = "ht!"
)
}
\arguments{
\item{pth}{\itemize{
\item path to image relative to \code{root} of document to be created
}}

\item{cap}{\itemize{
\item the caption for the image
}}

\item{lbl}{\itemize{
\item the latex label for the image
}}

\item{width}{\itemize{
\item figure width (in inches)
}}

\item{height}{\itemize{
\item figure height (in inches)
}}

\item{align}{\itemize{
\item character string indicating latex alignment
}}
}
\value{
nothing.
}
\description{
Function create and paste the latex code to insert an image into a latex document.
}
\details{
Intended for use in a QMD rendered in a latex environment.
The necessary latex code to insert an image into, e.g., a pdf document
is created and \link{cat}'ed. The QMD chunk that employs this function should
set \code{results="asis"} in the chunk header information.

The corresponding function for insertion into HTML is \link{insertImageIntoHTML}.
}
