% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertKblIntoQMD.R
\name{insertKblIntoQMD}
\alias{insertKblIntoQMD}
\title{Insert (print) a \link{knitr} \code{kable} object into a QMD (Quarto markdown) document}
\usage{
insertKblIntoQMD(kbl, lbl, cap = lbl, ori = "L")
}
\arguments{
\item{kbl}{\itemize{
\item \link{knitr} \code{kable} object
}}

\item{lbl}{\itemize{
\item table label (i.e., of the form "tbl-..." for cross-references)
}}

\item{cap}{\itemize{
\item table caption
}}

\item{ori}{\itemize{
\item orientation ("L" or "P")
}}
}
\value{
list if output environment is latex, otherwise NULL
}
\description{
Function to insert (print) a \link{knitr} \code{kable} object into a QMD (Quarto markdown) document.
}
\details{
This resulting table will be
\itemize{
\item{printed "here" in the text if the output environment is html}
\item{printed in the Tables section in the text if the output environment is latex}
}

If the output environment is latex, then a list to add to a \code{lstTbls}
list is returned, otherwise an invisible NULL is returned.

If the output environment is not latex, a QMD cell with table caption is printed via \link{cat}
to the output environment (typically an intermediate markdown file used in the
process of creating the final output).
}
