% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFunctions.R
\name{isOutputHTML}
\alias{isOutputHTML}
\title{Check if output is html}
\usage{
isOutputHTML()
}
\value{
TRUE/FALSE
}
\description{
Check if output is html
}
\details{
Checks to see if output is html in a \link{knitr} rendering context.
}
