% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFunctions.R
\name{isOutputPDF}
\alias{isOutputPDF}
\title{Check if output is pdf/latex}
\usage{
isOutputPDF()
}
\value{
TRUE/FALSE
}
\description{
Check if output is pdf/latex
}
\details{
Checks to see if output is pdf/latex in a \link{knitr} rendering context.
}
