% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberFunctions.R
\name{last}
\alias{last}
\title{Get the last n values of a vector}
\usage{
last(x, n = 1)
}
\arguments{
\item{x}{\itemize{
\item vector
}}

\item{n}{\itemize{
\item number of last values to get (default is 1)
}}
}
\value{
a vector of length \code{n}
}
\description{
Function to get the last \code{n} values of a vector.
}
\details{
none.
}
