% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{numKG2T}
\alias{numKG2T}
\title{Apply a pretty format to numbers in kg, converted to t}
\usage{
numKG2T(x)
}
\arguments{
\item{x}{\itemize{
\item numeric or character vector in kg
}}
}
\value{
\itemize{
\item character vector in t, formatted using \code{\link[=numT]{numT()}}
}
}
\description{
Function to apply a pretty format to numbers in kg, converted to t.
}
\details{
\itemize{
\item Uses \code{\link[=stripNumSyms]{stripNumSyms()}} (if necessary) and \code{\link[=numT]{numT()}}.
}
}
