% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{numKTinT}
\alias{numKTinT}
\title{Apply a pretty format to numbers in 1000s t, converted to t}
\usage{
numKTinT(x)
}
\arguments{
\item{x}{\itemize{
\item numeric or character vector in 1000s t
}}
}
\value{
\itemize{
\item character vector in t, formatted using \code{\link[=numT]{numT()}}
}
}
\description{
Function to apply a pretty format to numbers in 1000s, converted to t.
}
\details{
\itemize{
\item Uses \code{\link[=stripNumSyms]{stripNumSyms()}} (if necessary) and \code{\link[=numT]{numT()}}.
}
}
