% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{numTinKT}
\alias{numTinKT}
\title{Apply a pretty format to numbers in t, converted to 1000s t}
\usage{
numTinKT(x)
}
\arguments{
\item{x}{\itemize{
\item numeric or character vector in t
}}
}
\value{
\itemize{
\item character vector in 1000s t, formatted using \code{\link[=numKT]{numKT()}}
}
}
\description{
Function to apply a pretty format to numbers in t, converted to 1000s t.
}
\details{
\itemize{
\item Uses \code{\link[=stripNumSyms]{stripNumSyms()}} (if necessary) and \code{\link[=numKT]{numKT()}}.
}
}
