% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printGGplot.R
\name{printGGplot}
\alias{printGGplot}
\title{"Print" ggplot2 figure in a Quarto markdown document}
\usage{
printGGplot(
  p,
  lbl = NULL,
  pth = NULL,
  cap = NULL,
  asp = NULL,
  wid = NULL,
  dpi = NULL,
  xtraLbl = NULL,
  xtraFigFN = xtraLbl,
  xtraCap = NULL
)
}
\arguments{
\item{p}{\itemize{
\item a ggplot2 figure object
}}

\item{lbl}{\itemize{
\item figure label (OPTIONAL; should start with "fig-" if given)
}}

\item{pth}{\itemize{
\item path to saved figure (OPTIONAL)
}}

\item{cap}{\itemize{
\item figure caption (OPTIONAL)
}}

\item{asp}{\itemize{
\item figure aspect (height/width) (OPTIONAL)
}}

\item{wid}{\itemize{
\item figure width (inches) (OPTIONAL)
}}

\item{dpi}{\itemize{
\item dots-per-inch (OPTIONAL)
}}

\item{xtraLbl}{\itemize{
\item extra string to add to label (helpful for printing multiple plots in the same chunk) (OPTIONAL)
}}

\item{xtraFigFN}{\itemize{
\item extra string to add to path to saved figure (defaults to \code{xtraLbl}) (OPTIONAL)
}}

\item{xtraCap}{\itemize{
\item extra string to add to caption. (OPTIONAL)
}}
}
\value{
NULL, if the \link{knitr} environment is not latex. Otherwise, a lstFigs list that should be
appended to the QMD's lstFigs list.
}
\description{
Function to "print" ggplot2 figure in a Quarto markdown document.
}
\details{
All inputs except \code{p}, the plot to print/save, are optional. Values for \code{lbl}, \code{pth}, \code{cap}
\code{asp}, \code{wid}, and \code{dpi} that are not specified by the user will be taken from \link{knitr}'s
current chunk options.
}
