% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacementFunctions.R
\name{replaceNAs}
\alias{replaceNAs}
\title{Replace NAs in character vector with something else}
\usage{
replaceNAs(x, rpl = "--")
}
\arguments{
\item{x}{\itemize{
\item character vector
}}

\item{rpl}{\itemize{
\item replacement for NAs
}}
}
\value{
a character vector with NAs replaced
}
\description{
Function to replace NAs in character vector with something else.
}
\details{
None.
}
