% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeLabels.R
\name{sanitizeLabels}
\alias{sanitizeLabels}
\title{Sanitize labels by replacing illegal characters with "-"}
\usage{
sanitizeLabels(x, env = "latex")
}
\arguments{
\item{x}{\itemize{
\item character vector to sanitize
}}

\item{env}{\itemize{
\item environment for sanitizing characters (default="latex")
}}
}
\value{
character vector with sanitized labels.
}
\description{
Function to sanitize labels by replacing illegal characters with "-".
}
\details{
Currently sanitizes the following characters: "&", "\%", "_" in
a latex environment. Note that this function provides different results
than \code{\link[wtsMarkdown:sanitizeLabels]{wtsMarkdown::sanitizeLabels()}}.
}
