% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{stripNumSyms}
\alias{stripNumSyms}
\title{Strip number symbols ("$",",") from a number}
\usage{
stripNumSyms(x)
}
\arguments{
\item{x}{\itemize{
\item (character or numeric) vector to strip symbols from
}}
}
\value{
numeric vector.
}
\description{
Function to strip number symbols ("$",",") from a number.
}
\details{
Strips dollar symbols and comma delineators from a vector of formatted numbers.
Simply passes \code{x} through if it is already numeric.
}
