% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tblr_to_kbl.R
\name{convert_tblr_to_kbl}
\alias{convert_tblr_to_kbl}
\title{Convert \link{tables} \code{tabular} object to \link{knitr} \code{kable} object}
\usage{
convert_tblr_to_kbl(tblr, col_spec, isHTM)
}
\arguments{
\item{tblr}{\itemize{
\item \link{tables} \code{tabular} object to convert
}}

\item{col_spec}{\itemize{
\item vector of column ids to apply "border_right" to using \code{\link[kableExtra:column_spec]{kableExtra::column_spec()}}
}}

\item{isHTM}{\itemize{
\item flag indicating whether output is for html (or pdf)
}}
}
\value{
\link{knitr} \code{kable} object
}
\description{
Function to convert \link{tables} \code{tabular} object to \link{knitr} \code{kable} object.
}
\details{
TBD.
}
