% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{fmtT}
\alias{fmtT}
\title{Format numeric 'x' to specified number of digits}
\usage{
fmtT(x, d = 0)
}
\arguments{
\item{x}{\itemize{
\item numeric vector
}}

\item{d}{\itemize{
\item number of significant digits for output (default = 0)
}}
}
\value{
character vector with formatted values
}
\description{
Function to format numeric 'x' to specified number of digits.
}
\details{
Shortcut to \code{\link[base:formatc]{base::formatC()}} function using format='f', big.mark=',',digits=\code{d}
}
