% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getLabel}
\alias{getLabel}
\title{Create a Quarto/knitr chunk label}
\usage{
getLabel(xtra = NULL)
}
\arguments{
\item{xtra}{\itemize{
\item extra string toadd to default label
}}
}
\value{
a string
}
\description{
Function to create a Quarto/knitr chunk label.
}
\details{
If the function is not run within the context of a knitr chunk,
then 'lbl' is used as the default label.

If the chunk label starts with "tbl_" or "fig_", it is assumed that the
associated chunk output is a table or figure that should be indexed and
cross-referenced, so the "_" is replaced by "-" in order that Quarto/Pandoc's
automatic indexing tracks the output correctly.

Non-alphanumeric characters in \code{xtra}
are replaced by "-" in order to conform to Quarto labels.
}
