% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFunctions.R
\name{num}
\alias{num}
\title{Pretty format for a vector of numbers}
\usage{
num(x, d = 0)
}
\arguments{
\item{x}{\itemize{
\item vector (numeric or character) of numbers to apply pretty format to
}}

\item{d}{\itemize{
\item number of significant digits to keep
}}
}
\value{
character vector
}
\description{
Function to apply a pretty format to number
}
\details{
applies (base)\link{format} with format=f, big.mark=",",digits=\code{d}.
NAs are represented as "--".
}
